Rename file back to original to avoid possible problems in the future
