+++ src/__test__/components/data-exploration/differential-expression-tool/DiffExprManager-regression-load-on-second-render.test.jsx
@@ -17,7 +17,7 @@ import { DIFF_EXPR_LOADING, DIFF_EXPR_LOADED } from '../../../../redux/actionTyp
 jest.mock('localforage');
 
 // ensure isBrowser is `true`
-jest.mock('../../../../utils/Environment', () => ({
+jest.mock('../../../../utils/environment', () => ({
   __esModule: true,
   isBrowser: () => true,
 }));
