+++ src/__test__/components/data-exploration/differential-expression-tool/DiffExprManager.test.jsx
@@ -12,7 +12,7 @@ import DiffExprResults from '../../../../components/data-exploration/differentia
 import initialState from '../../../../redux/reducers/differentialExpression/initialState';
 
 jest.mock('localforage');
-jest.mock('../../../../utils/Environment', () => ({
+jest.mock('../../../../utils/environment', () => ({
   __esModule: true,
   isBrowser: () => false,
 }));
