+++ src/utils/cacheRequest.js
@@ -2,7 +2,7 @@
 import hash from 'object-hash';
 import cache from './cache';
 import sendWork from './sendWork';
-import { isBrowser } from './Environment';
+import { isBrowser } from './environment';
 import { calculateZScore } from './postRequestProcessing';
 
 const createObjectHash = (object) => hash.MD5(object);
